dom/base/test/test_bug902847.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=902847
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 902847</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  function testPaintextSerializerWithPlaceHolder() {

    const de = SpecialPowers.Ci.nsIDocumentEncoder;
    const Cc = SpecialPowers.Cc;

    // Create a plaintext encoder.
    var encoder = Cc["@mozilla.org/layout/documentEncoder;1?type=text/plain"]
                  .createInstance(de);
    var flags = de.OutputRaw |
                de.OutputLFLineBreak |
                de.OutputDontRemoveLineEndingSpaces;
    encoder.init(document, "text/plain", flags);

    function toPlaintext(id) {
      var element = document.getElementById(id);
      var range = document.createRange();
      range.selectNodeContents(element);
      encoder.setRange(range);
      return encoder.encodeToString().replace('\n', '\\n', 'g')
                                     .replace('\r', '\\r', 'g');
    }

    // Test cases.
    is(toPlaintext("case1"), "Hello \\nboy!", "Case 1 failed.");
    is(toPlaintext("case2"), "Hello \\nboy!", "Case 2 failed.");
    is(toPlaintext("case3"), "Hello \\nboy!", "Case 3 failed.");
    is(toPlaintext("case4"), "Hello  \\nboy!", "Case 4 failed.");
    SimpleTest.finish();
  }

  addLoadEvent(testPaintextSerializerWithPlaceHolder);
  SimpleTest.waitForExplicitFinish();
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=902847">Mozilla Bug 902847</a>
<p id="display"></p>
<div id="content">
  <p id="case1">Hello <br>boy!</p>
  <p id="case2">Hello  <br>boy!</p>
  <p id="case3">Hello&nbsp;<br>boy!</p>
  <p id="case4">Hello&nbsp;&nbsp;<br>boy!</p>
</div>
<pre id="test">
</pre>
</body>
</html>
